requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
idna==2.10
nums-from-string==0.1.2
python-slugify==4.0.1
requests==2.25.0
soupsieve==2.0.1
text-unidecode==1.3
urllib3==1.26.2
